+++ CyRK/cy/cysolver.hpp
@@ -7,8 +7,6 @@
 #include <memory>
 
 #include "common.hpp"
-#include "cysolution.hpp"
-#include "dense.hpp"
 #include "cy_array.hpp"
 
 // !!!
@@ -23,8 +21,25 @@
 #include <Python.h>
 #include "pysolver_cyhook_api.h"
 
+// We need to forward declare the CySolverResult so that the solver can make calls to its methods
+class CySolverResult;
+class CySolverDense;
+
 struct _object;
 typedef _object PyObject;
+
+struct NowStatePointers 
+{
+    double* t_now_ptr;
+    double* y_now_ptr;
+    double* dy_now_ptr;
+    
+    NowStatePointers():
+        t_now_ptr(nullptr), y_now_ptr(nullptr), dy_now_ptr(nullptr) { }
+
+    NowStatePointers(double* t_now_ptr, double* y_now_ptr, double* dy_now_ptr):
+        t_now_ptr(t_now_ptr), y_now_ptr(y_now_ptr), dy_now_ptr(dy_now_ptr) { }
+};
 
 class CySolverBase {
 
@@ -32,8 +47,6 @@ class CySolverBase {
 protected:
     virtual void p_estimate_error();
     virtual void p_step_implementation();
-    virtual CySolverDense* p_dense_output_heap();
-    virtual void p_dense_output_stack(CySolverDense& dense_output_ptr);
 
 public:
     CySolverBase();
@@ -41,7 +54,7 @@ public:
     CySolverBase(
         // Input variables
         DiffeqFuncType diffeq_ptr,
-        std::shared_ptr<CySolverResult> const storage_ptr,
+        std::shared_ptr<CySolverResult> storage_sptr,
         const double t_start,
         const double t_end,
         const double* const y0_ptr,
@@ -56,15 +69,18 @@ public:
         PreEvalFunc pre_eval_func = nullptr
     );
 
-    void change_storage(std::shared_ptr<CySolverResult> new_storage_ptr, bool auto_reset = true);
     bool check_status() const;
     virtual void reset();
+    void offload_to_temp();
+    void load_back_from_temp();
+    virtual void set_Q_array(double* Q_ptr, unsigned int* Q_order_ptr);
     virtual void calc_first_step_size();
     void take_step();
     void solve();
     // Diffeq can either be the C++ class method or the python hook diffeq. By default set to C++ version.
     void cy_diffeq() noexcept;
     std::function<void(CySolverBase*)> diffeq;
+    NowStatePointers get_now_state();
 
     // PySolver methods
     void set_cython_extension_instance(PyObject* cython_extension_class_instance, DiffeqMethod py_diffeq_method);
@@ -75,9 +91,7 @@ protected:
     // ** Attributes **
 
     // Time variables
-    double t_start       = 0.0;
-    double t_end         = 0.0;
-    double t_old         = 0.0;
+    double t_tmp         = 0.0;
     double t_delta       = 0.0;
     double t_delta_abs   = 0.0;
     double direction_inf = 0.0;
@@ -85,18 +99,6 @@ protected:
     // Dependent variables
     double num_y_dbl  = 0.0;
     double num_y_sqrt = 0.0;
-    // The size of the stack allocated tracking arrays is equal to the maximum allowed `num_y` (25).
-    double y0[Y_LIMIT]    = { 0.0 };
-    double y_old[Y_LIMIT] = { 0.0 };
-    double y_now[Y_LIMIT] = { 0.0 };
-    // For dy, both the dy/dt and any extra outputs are stored. So the maximum size is `num_y` (25) + `num_extra` (25)
-    double dy_old[DY_LIMIT] = { 0.0 };
-    double dy_now[DY_LIMIT] = { 0.0 };
-
-    // dy_now_ptr and y_now_ptr are declared in public.
-    double* const y0_ptr     = &y0[0];
-    double* const y_old_ptr  = &y_old[0];
-    double* const dy_old_ptr = &dy_old[0];
 
     // Integration step information
     size_t max_num_steps = 0;
@@ -110,10 +112,8 @@ protected:
     double* t_eval_ptr       = t_eval_vec.data();
     size_t t_eval_index_old  = 0;
     size_t len_t_eval        = 0;
+    bool skip_t_eval         = false;
     bool use_t_eval          = false;
-
-    // Information on capturing extra information during integration.
-    int num_extra = 0;
 
     // Function to send to diffeq which is called before dy is calculated
     PreEvalFunc pre_eval_func = nullptr;
@@ -136,19 +136,30 @@ public:
 
     // Status attributes
     int status = -999;
-    int integration_method = -1;
+    unsigned int integration_method = 999;
 
     // Meta data
-    unsigned int num_dy = 0;
-    unsigned int num_y  = 0;
+    unsigned int num_dy    = 0;
+    unsigned int num_y     = 0;
+    unsigned int num_extra = 0;
+
+    // The size of the stack allocated tracking arrays is equal to the maximum allowed `num_y` (25).
+    double y0[Y_LIMIT]    = { 0.0 };
+    double y_old[Y_LIMIT] = { 0.0 };
+    double y_now[Y_LIMIT] = { 0.0 };
+    double y_tmp[Y_LIMIT] = { 0.0 };
+    // For dy, both the dy/dt and any extra outputs are stored. So the maximum size is `num_y` (25) + `num_extra` (25)
+    double dy_old[DY_LIMIT] = { 0.0 };
+    double dy_now[DY_LIMIT] = { 0.0 };
+    double dy_tmp[DY_LIMIT] = { 0.0 };
 
     // Result storage
-    std::shared_ptr<CySolverResult> storage_ptr = std::make_shared<CySolverResult>();
+    std::shared_ptr<CySolverResult> storage_sptr = nullptr;
 
     // State attributes
     size_t len_t       = 0;
-    double t_now[1]    = {0.0};
-    double* const t_now_ptr  = &t_now[0];
-    double* const y_now_ptr  = &y_now[0];
-    double* const dy_now_ptr = &dy_now[0];
+    double t_now       = 0.0;
+    double t_start     = 0.0;
+    double t_end       = 0.0;
+    double t_old       = 0.0;
 };