+++ CyRK/cy/pysolver.pyx
@@ -5,10 +5,11 @@ from libc.string cimport memcpy
 from libcpp.cmath cimport fmin, fabs
 
 from CyRK.utils.memory cimport make_shared
-from CyRK.cy.cysolver_api cimport find_expected_size, WrapCySolverResult, INF, EPS_100, Y_LIMIT, DY_LIMIT
+from CyRK.cy.cysolver_api cimport find_expected_size, WrapCySolverResult, NowStatePointers, INF, EPS_100, Y_LIMIT, DY_LIMIT
 
 import numpy as np
-np.import_array()
+cimport numpy as cnp
+cnp.import_array()
 
 # =====================================================================================================================
 # PySolver Class (holds the intergrator class and reference to the python diffeq function)
@@ -46,11 +47,22 @@ cdef class WrapPyDiffeq:
             self.pass_dy_as_arg = True
         else:
             self.pass_dy_as_arg = False
+        
+        # The cysolver instance is set after inititalization via `set_state`. For now it is null
+        # self.solver_sptr.reset()
     
-    cdef void set_state(self, double* dy_ptr, double* t_ptr, double* y_ptr) noexcept:
-        self.dy_now_ptr = dy_ptr
-        self.t_now_ptr  = t_ptr
-        self.y_now_ptr  = y_ptr
+    def __dealloc__(self):
+        # Reset the solver shared pointer
+        if self.solver_sptr:
+            self.solver_sptr.reset()
+    
+    cdef void set_state(self, shared_ptr[CySolverBase] solver_sptr) noexcept:
+        self.solver_sptr = solver_sptr
+        cdef NowStatePointers solver_state = self.solver_sptr.get().get_now_state()
+        
+        self.t_now_ptr  = solver_state.t_now_ptr
+        self.y_now_ptr  = solver_state.y_now_ptr
+        self.dy_now_ptr = solver_state.dy_now_ptr
 
         # Create memoryviews of the pointers
         self.y_now_view  = <double[:self.num_y]>self.y_now_ptr
@@ -60,17 +72,17 @@ cdef class WrapPyDiffeq:
         # That is why we use `PyArray_SimpleNewFromData` instead of a more simple `asarray`.
         # Note that it is not safe to return these arrays outside of this class because they may get deallocated while
         # the numpy array still points to the underlying memory.
-        cdef np.npy_intp[1] shape
-        cdef np.npy_intp* shape_ptr = &shape[0]
-        shape_ptr[0] = <np.npy_intp>self.num_y
+        cdef cnp.npy_intp[1] shape
+        cdef cnp.npy_intp* shape_ptr = &shape[0]
+        shape_ptr[0] = <cnp.npy_intp>self.num_y
         
-        self.y_now_arr = np.PyArray_SimpleNewFromData(1, shape_ptr, np.NPY_DOUBLE, self.y_now_ptr)
+        self.y_now_arr = cnp.PyArray_SimpleNewFromData(1, shape_ptr, cnp.NPY_DOUBLE, self.y_now_ptr)
         
         # Do the same for dy if the user provided the appropriate kind of differential equation.
         if self.pass_dy_as_arg:
             self.dy_now_view = <double[:self.num_dy]>self.dy_now_ptr
-            shape[0]         = <np.npy_intp>self.num_dy  # dy may have a larger shape than y
-            self.dy_now_arr  = np.PyArray_SimpleNewFromData(1, shape_ptr, np.NPY_DOUBLE, self.dy_now_ptr)   
+            shape[0]         = <cnp.npy_intp>self.num_dy  # dy may have a larger shape than y
+            self.dy_now_arr  = cnp.PyArray_SimpleNewFromData(1, shape_ptr, cnp.NPY_DOUBLE, self.dy_now_ptr)   
 
     cdef void diffeq(self) noexcept:
         # Run python diffeq
@@ -207,7 +219,7 @@ def pysolve_ivp(
         raise AttributeError("Maximum step size must be a postive float.")
 
     # Build solution storage
-    cdef shared_ptr[CySolverResult] result_ptr = make_shared[CySolverResult](
+    cdef shared_ptr[CySolverResult] solution_sptr = make_shared[CySolverResult](
         num_y,
         num_extra,
         expected_size,
@@ -233,11 +245,12 @@ def pysolve_ivp(
             integration_method,
             <cpy_ref.PyObject*>diffeq_wrap,
             diffeq_func,
-            result_ptr,
+            solution_sptr,
             t_start,
             t_end,
             y0_ptr,
             num_y,
+            expected_size,
             num_extra,
             args_ptr,
             max_num_steps,
@@ -253,13 +266,8 @@ def pysolve_ivp(
             first_step
         )
 
-    # Get pointers to the solver's state variables so that the Python differential equation can use and update them.
-    cdef PySolverStatePointers state_pointers = solver.get_state_pointers()
-    diffeq_wrap.set_state(
-        state_pointers.dy_now_ptr,
-        state_pointers.t_now_ptr,
-        state_pointers.y_now_ptr
-        )
+    # Get pointers to the solver so that the Python differential equation can use and update its state variables (t_now, y_now, dy_now).
+    diffeq_wrap.set_state(solver.solution_sptr.get().solver_sptr)
     
     ##
     # Run the integrator!
@@ -268,7 +276,7 @@ def pysolve_ivp(
     
     # Wrap the solution in a python-safe wrapper class
     cdef WrapCySolverResult pyresult = WrapCySolverResult()
-    pyresult.set_cyresult_pointer(result_ptr)
+    pyresult.set_cyresult_pointer(solution_sptr)
 
     del solver